fix(dogfood): schedule completion of key changes
Previously, we incorrectly relied on `GetImuachainValidator` to
determine whether a consensus key event can be completed immediately or
scheduled for a later epoch.

If a validator is currently jailed, the validator object in x/dogfood
cannot be found; thus, the consensus key event would be processed
immediately. However, doing so would mean that any slashable actions
that took place prior to the jailing event would escape slashing. To
avoid that, the scheduling is always applied.