feat(evm): process x/group messages from authority
See #396 for the core rationale; this PR should be merged after that one

Since x/group's first group will (temporarily) act as the chain's
governance layer, why not have that governance layer be the contract
deployer and manager? This PR adds that capability as well as a unit
test demonstrating the ability to deploy and call owner-only functions.

Note the usage of `common.BytesToAddress` to act as the deployer; this
means that the longer Cosmos group address is cropped from the left and
retains the right-most 20 bytes in the Solidity context