refactor(delegation): avoid staker list
Marshaling and unmarshaling a large list is often a costly affair.
Instead of relying on a staker list, we use a sentinel value of
`[]byte{1}` to mark the presence of an operator to assetID to stakerID
delegation. Absence of this sentinel value indicates absence of such a
delegation.

The gas savings are enormous: adding a new staker to a list of size 500
costs 200k gas whereas marking a sentinel entry at that size is just
5180 gas